backend/requirements.txt
google-adk==1.27.0
fastapi==0.124.1
uvicorn[standard]==0.34.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.132.0
google-cloud-texttospeech>=2.16.0
google-cloud-storage>=2.16.0
python-dotenv==1.0.0
pydantic>=2.7.0
